@@ -34,33 +33,31 @@ app = "app.__main__:main"
 [tool.poetry.dependencies]
 alembic = "*"
 asyncpg = "*"
-click = "*"
-greenlet = "^1.1.3"
 hiredis = "*"
-httpx = "*"
 passlib = {version = "*", extras = ["argon2"]}
-pydantic = {extras = ["dotenv", "email"], version = "*"}
-python = ">=3.10,<4.0"
-redis = "*"
-rich = "*"
-sqlalchemy = {git = "https://github.com/sqlalchemy/sqlalchemy.git", branch = "main", extras = ["asyncio"]}
-starlite = {version = "^1.24.0", extras = ["brotli"]}
-starlite-jwt = "^1.4.0"
+pydantic = {version = ">=1.9.0", extras = [
+  "dotenv",
+  "typing_extensions",
+  "email",
+]}
+python = ">=3.11,<4.0"
+redis = ">=4.4.2"
+rich-click = ">=1.6.0"
+sqlalchemy = {version = ">=1.4.42", allow-prereleases = true}
+starlite = {version = ">=1.35", extras = ["standard", "jwt", 'cli', 'redis']}
+starlite-saqlalchemy = "*"
+starlite-users = {git = "https://github.com/LonelyVikingMichael/starlite-users.git", branch = "main"}
 uvicorn = {extras = ["standard"], version = "*"}
 
 [tool.poetry.group.dev.dependencies]
-bump2version = "*"
 coverage = {extras = ["toml"], version = "*"}
 cython = "*"
-hypothesis = {extras = ["cli"], version = "*"}
-ipykernel = "*"
-ipython = "*"
 pytest = "*"
 pytest-asyncio = "*"
 pytest-cov = "*"
+pytest-docker = "*"
 pytest-dotenv = "*"
 pytest-mock = "*"
-tox = "*"
 
 [tool.poetry.group.docs]
 optional = true
